Rename rbv2 -> yahpo_benchmark
